# specs/001-prd-app-only/contracts/api.yaml
openapi: 3.0.3
info:
  title: Yabaii - Japanese Price Comparison App API
  description: RESTful API for Japanese price comparison mobile application
  version: 1.0.0
  contact:
    name: API Team
    url: https://yabaii.day
  license:
    name: MIT
servers:
  - url: https://api.yabaii.day/v1
    description: Production
  - url: https://staging-api.yabaii.day/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /prices/{spuId}/history:
    get:
      tags: [Prices]
      summary: Get price history for a product
      operationId: getPriceHistory
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 90
        - name: platform
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Price history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistoryResponse'
  /prices/{spuId}/statistics:
    get:
      tags: [Prices]
      summary: Get price statistics for a product
      operationId: getPriceStatistics
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Price statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceStatisticsResponse'

  # Alert endpoints
  /alerts:
    get:
      tags: [Alerts]
      summary: Get user alerts
      operationId: getUserAlerts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAlertsResponse'

    post:
      tags: [Alerts]
      summary: Create new alert
      operationId: createAlert
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
  /alerts/{alertId}:
    delete:
      tags: [Alerts]
      summary: Delete alert
      operationId: deleteAlert
      security:
        - BearerAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Alert deleted

  # Deals endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    PriceHistoryResponse:
      type: object
      required:
        - history
        - statistics
      properties:
        history:
          type: array
          items:
            $ref: '#/components/schemas/PriceHistory'
        statistics:
          $ref: '#/components/schemas/PriceStatistics'

    # Request schemas
    CreateAlertRequest:
      type: object
      required:
        - productId
        - type
      properties:
        productId:
          type: string
        platformId:
          type: string
        type:
          type: string
          enum: [price_drop, price_low, in_stock, any_change]
        condition:
          $ref: '#/components/schemas/AlertCondition'
...
    PriceHistory:
      type: object
      required:
        - productId
        - platformId
        - price
        - recordedAt
      properties:
        productId:
          type: string
        platformId:
          type: string
        price:
          type: number
        currency:
          type: string
        recordedAt:
          type: string
          format: date-time

    # Additional schemas for completeness
...
    AlertCondition:
      type: object
      properties:
        threshold:
          type: number
        percentage:
          type: number
        lowestHistorical:
          type: boolean
        platformIds:
          type: array
          items:
            type: string
...
    PriceStatistics:
      type: object
      properties:
        currentPrice:
          type: number
        averagePrice:
          type: number
        lowestPrice:
          type: number
        highestPrice:
          type: number
        priceChangePercent:
          type: number
        trend:
          type: string
          enum: [rising, falling, stable]

    # Request/Response schemas for remaining endpoints
...
    UserAlertsResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
    AlertResponse:
      type: object
      properties:
        alert:
          $ref: '#/components/schemas/Alert'
    Alert:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        platformId:
          type: string
        type:
          type: string
        condition:
          $ref: '#/components/schemas/AlertCondition'
        active:
          type: boolean
        triggered:
          type: boolean
        createdAt:
          type: string
          format: date-time
        triggeredAt:
          type: string
          format: date-time
...